youtube-scraper-api/requirements.txt
requests
yt-dlp
webvtt-py
fastapi
uvicorn